requirements.txt
# Core Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
starlette==0.27.0

# Authentication & Security
PyJWT==2.8.0
itsdangerous==2.1.2
secrets-python==3.3.0

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0
redis==5.0.1
elasticsearch==8.11.0

# Machine Learning & AI
scikit-learn==1.3.2
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4
joblib==1.3.2
torch==2.1.0
xgboost==2.0.1

# Network Security & Monitoring
scapy==2.5.0
psutil==5.9.6
netaddr==0.10.1
python-nmap==0.7.1
dnspython==2.4.2

# Cryptography & Security
cryptography==41.0.8
bcrypt==4.1.2
pycryptodome==3.19.0

# Threat Intelligence & Honeypots
yara-python==4.3.1
stix2==3.0.1
taxii2-client==2.3.0
twisted==23.10.0
paramiko==3.4.0

# Data Visualization & Reporting
plotly==5.17.0
dash==2.14.2
matplotlib==3.8.2
seaborn==0.13.0
fpdf2==2.7.6

# Async Operations & Networking
aiofiles==23.2.1
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Configuration & Logging
pyyaml==6.0.1
python-dotenv==1.0.0
structlog==23.2.0
loguru==0.7.2

# System Utilities
schedule==1.2.0
watchdog==3.0.0
click==8.1.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
isort==5.12.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Monitoring & Metrics
prometheus-client==0.19.0

# Container Orchestration (Optional)
docker==6.1.3
kubernetes==28.1.0

# Additional Security Tools
nmap==0.0.1
pypcap==1.2.3
